Remove Project::Queue::Base and subclasses
Massively cleans up and reduces complexity around project queues.

Now that queues just operate on an `ActiveRecord::Relation` we can get
rid of the subclasses for each queue type and replace them with factory
methods that set up a named queue, backed by the session, with the
correct requests.

This also allows `Project::Queue` to become a class rather than a
containing module, which makes for a much nicer language throughout the
app. It also reduces all the spec duplication – we now just have a set
of specs for general queue behaviour, and then test the factory methods
create a queue with the particular attributes we're interested in.
