+++ app/controllers/projects/extracts_controller.rb
@@ -1,4 +1,4 @@
-require_dependency 'project/queue/extractable'
+require_dependency 'project/queue'
 
 # Extract data from a Project
 class Projects::ExtractsController < Projects::BaseController
@@ -27,9 +27,7 @@ class Projects::ExtractsController < Projects::BaseController
   def update
     authorize! :read, @project
 
-    backend =
-      Project::Queue::SessionBackend.primed(session, @project, :extractable)
-    queue = Project::Queue::Extractable.new(@project.info_requests, backend)
+    queue = Project::Queue.extractable(@project, session)
     queue.skip(@info_request)
 
     redirect_to project_extract_path(@project), notice: _('Skipped!')
@@ -66,9 +64,7 @@ class Projects::ExtractsController < Projects::BaseController
         url_title: params[:url_title]
       )
     else
-      backend =
-        Project::Queue::SessionBackend.primed(session, @project, :extractable)
-      @queue = Project::Queue::Extractable.new(@project.info_requests, backend)
+      @queue = Project::Queue.extractable(@project, session)
       @info_request = @queue.next
     end
   end
