--- app/models/project/queue/extractable.rb
@@ -1,8 +1,0 @@
-module Project::Queue
-  # Public: Extractable requests in the given Project for the given User.
-  class Extractable < Base
-    def initialize(info_requests, backend)
-      super(info_requests.extractable, backend)
-    end
-  end
-end
